+++ cmd/validate.go
@@ -88,5 +88,5 @@ var validateCmd = &cobra.Command{
 }
 
 func init() {
-	rootCmd.AddCommand(validateCmd)
+	typeCmd.AddCommand(validateCmd)
 }